Refactored the code according to the comments suggested.
