+++ src/libs/ui/src/molecules/SearchDailogContent/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+interface ModalBackgroundProps {
+    onClick: () => void;
+}
+
+const ModalBackground: React.FC<ModalBackgroundProps> = ({ onClick }) => (
+  <div className="absolute inset-0" onClick={onClick}></div>
+);
+
+export default ModalBackground;
