--- src/pages/PlpPage/index.tsx
@@ -1,10 +1,0 @@
-// import PlpPageTemplate from '@ui/templates/PlpPage'
-// import React from 'react'
-
-// function PlpPage() {
-//   return (
-//     <PlpPageTemplate></PlpPageTemplate>
-//   )
-// }
-
-// export default PlpPage
